+++ tiny/net/client.cpp
@@ -80,7 +80,7 @@ bool Client::listen(const double &dt)
     {
         receivedData = false;
         
-        if (SDLNet_CheckSockets(selector, 1000.0*dt) > 0)
+        if (SDLNet_CheckSockets(selector, static_cast<Uint32>(1000.0*dt)) > 0)
         {
             if (SDLNet_SocketReady(socket) != 0)
             {
